Change CCL input precision from long to int8 to improve performance

+++ QEfficient/transformers/models/llama4/modeling_llama4.py
@@ -1225,7 +1225,7 @@ class QEffLlama4ForConditionalGeneration(Llama4ForConditionalGeneration):
             lang_inputs["batch_index"] = torch.arange(bs).view(bs, 1)
 
         if comp_ctx_lengths is not None:
-            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.long)
+            lang_inputs["comp_ctx_lengths"] = torch.randint(0, 100, (40,), dtype=torch.int8)
 
         inputs = {}
         if kv_offload:
